Conflicts:
	src/whelks-core/.gitignore
	src/whelks-core/build.gradle
	src/whelks-core/src/main/groovy/whelks/Converters.groovy
	src/whelks-core/src/main/groovy/whelks/ElasticSearch.groovy
	src/whelks-core/src/main/groovy/whelks/RestManager.groovy
	src/whelks-core/src/main/groovy/whelks/Tools.groovy
	src/whelks-core/src/main/groovy/whelks/Whelk.groovy
	src/whelks-core/src/main/java/se/kb/libris/whelks/Document.java
	src/whelks-core/src/main/java/se/kb/libris/whelks/SearchResult.java
	src/whelks-core/src/main/java/se/kb/libris/whelks/Whelk.java
	src/whelks-core/src/main/java/se/kb/libris/whelks/WhelkManager.java
	src/whelks-core/src/main/java/se/kb/libris/whelks/api/API.java
	src/whelks-core/src/main/java/se/kb/libris/whelks/basic/BasicDocument.java
	src/whelks-core/src/main/java/se/kb/libris/whelks/basic/BasicWhelk.java
	src/whelks-core/src/main/java/se/kb/libris/whelks/component/Component.java
	src/whelks-core/src/main/java/se/kb/libris/whelks/component/Index.java
	src/whelks-core/src/main/java/se/kb/libris/whelks/component/QuadStore.java
	src/whelks-core/src/main/java/se/kb/libris/whelks/component/Storage.java
	src/whelks-core/src/main/java/se/kb/libris/whelks/plugin/Plugin.java
	src/whelks-core/src/main/java/se/kb/libris/whelks/plugin/Trigger.java
	src/whelks-core/src/main/java/se/kb/libris/whelks/storage/Storage.java
	src/whelks-core/src/main/java/se/kb/libris/whelks/test/SerialiseTest.java
	src/whelks-core/src/main/java/se/kb/libris/whelks/test/TestWhelk.java
	src/whelks-core/src/main/resources/sug_json.py
	src/whelks-core/src/test/groovy/libris/whelks/WhelksTest.groovy
	src/whelks-import/build.gradle

+++ src/whelks-core/src/main/java/se/kb/libris/whelks/SearchResult.java
@@ -1,7 +1,8 @@
 package se.kb.libris.whelks;
 
 public interface SearchResult<T extends Document> {
-    int getNumberOfHits();
+
+    long getNumberOfHits();
     Iterable<Facet> getFacets();
     Iterable<Hit<T>> getHits();
